Fix logic for "for..in/of" loops
The code currently throws the error

TypeError: Cannot read property 'loc' of undefined

That's because `for...in/of` loops don't have a "test" property
(http://astexplorer.net/#/CUhuGjRDKX), but have "left" and "right"
instead. It seems the line number is only used for error reporting so
this fix simply uses "node.test" if it exists, otherwise "node.left".

Added a simple test with a "for...in" loop just to verify that it is
transformed without error.

Test are passing.